Move broadcast calendar helpers to their own module
